requirements.txt
altair==5.4.1
asgiref==3.8.1
attrs==24.2.0
bizerror==0.4.11
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
Cython==3.0.11
dictop==0.2.2
Django==4.2
django-app-requires==0.3.4
django-click==2.4.0
django-cors-headers==4.4.0
django-object-toolbar-admin==0.1.2
django-redis-data-manager==0.1.8
django-safe-fields==0.2.2
django-server-controller==0.5.1
django-static-fontawesome==6.5.2.0
django-static-jquery3==6.3.7.1
djangorestframework==3.14.0
drf-spectacular==0.27.2
et-xmlfile==1.1.0
fastutils==0.43.4
future==1.0.0
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.43
gunicorn==23.0.0
h11==0.14.0
idna==3.8
inflection==0.5.1
Jinja2==3.1.4
jsonformatter==0.3.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
magic-import==0.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
narwhals==1.5.5
numpy==2.1.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
ply==3.11
protobuf==5.28.0
psutil==6.0.0
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.0
pycryptodome==3.20.0
pydeck==0.9.1
Pygments==2.18.0
pypinyin==0.52.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzk==0.9
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
selenium==4.24.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
streamlit==1.38.0
tenacity==8.5.0
thriftpy2==0.5.2
toml==0.10.2
tornado==6.4.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.2
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
zenutils==0.5.5
zk==0.8.5
